Decode/encode on Task.detached
